requirements.txt
django==2.1.7
djangorestframework==3.9.2
django-cors-headers
